+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### release 1.6.0 notes
+add select block fucntion. shortkey(**alt + x**) <br>
+now support function, html tag, if, for, while, json, array block select
+![](./examples/selectBlock.gif)
+
 ### release 1.5.1 notes
 fix bug jump to definition with postfix
 
@@ -6,7 +11,7 @@ exchange rem to px or exchange px to rem for all file through command
 ![](./examples/rem-file.gif)
 
 ### release 1.4.5 notes
-rem px exchange
+rem px exchange, shortkey (**alt + z**)
 ![](./examples/rem.gif)
 rem px exchange setting
 ![](./examples/rem-config.png)